src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
astropy=5.0.3=py38hce1f21e_0
astroquery=0.4.6=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
bcrypt=3.2.0=py38h7b6447c_0
beautifulsoup4=4.10.0=pyh06a4308_0
black=19.10b0=py_0
blas=1.0=mkl
boost=1.74.0=py38h2b96118_5
boost-cpp=1.74.0=h312852a_4
bottleneck=1.3.4=py38hce1f21e_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.3.29=h06a4308_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cryptography=3.4.8=py38hd23ed53_0
cspice=67=h166bdaf_1
dbus=1.13.18=hb2f20db_0
dill=0.3.4=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
glib=2.69.1=h4ff587b_1
globus-sdk=3.6.0=pyhd8ed1ab_0
greenlet=1.1.1=py38h295c915_0
gsoap_abi=2.8.110=ha770c72_0
htcondor-classads=9.6.0=h9d75d03_0
html5lib=1.1=pyhd3eb1b0_0
humanize=3.10.0=pyhd3eb1b0_0
icu=68.1=h2531618_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jeepney=0.7.1=pyhd3eb1b0_0
keyring=23.4.0=py38h06a4308_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
ld_impl_linux-64=2.35.1=h7274673_9
libcondor_utils=9.6.0=h35997f9_0
libcurl=7.80.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1d223b6_14
libnghttp2=1.46.0=hce63b2e_0
libpq=12.9=hc2c5182_1
libsodium=1.0.18=h7b6447c_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=11.2.0=he4da1e4_14
libuuid=2.32.1=h7f98852_1000
lz4-c=1.9.3=h295c915_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munge=0.5.13=h4de3113_1
mypy_extensions=0.4.3=py38h06a4308_1
ncurses=6.3=h7f8727e_2
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py38h295c915_1
paramiko=2.8.1=pyhd3eb1b0_0
parsl=1.2.0=pyhd8ed1ab_0
pathspec=0.7.0=py_0
pcre=8.45=h295c915_0
pip=21.2.4=py38h06a4308_0
psutil=5.8.0=py38h27cfd23_1
psycopg2=2.8.6=py38h3c74f83_1
psycopg2-binary=2.8.6=pyhd8ed1ab_3
pycparser=2.21=pyhd3eb1b0_0
pyerfa=2.0.0=py38h27cfd23_0
pyjwt=2.1.0=py38h06a4308_0
pynacl=1.4.0=py38h7b6447c_1
pyopenssl=19.1.0=py38_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-htcondor=9.6.0=py38hda66bd9_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pyvo=1.3=pyhd8ed1ab_0
pyyaml=6.0=py38h7f8727e_1
pyzmq=22.3.0=py38h295c915_2
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py38h7f8727e_0
requests=2.27.1=pyhd3eb1b0_0
scipy=1.7.3=py38hc147768_0
scitokens-cpp=0.6.3=h00015f2_2
secretstorage=3.3.1=py38h06a4308_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
spiceypy=5.0.1=pyhd8ed1ab_0
sqlalchemy=1.4.32=py38h7f8727e_0
sqlite=3.38.2=hc218d9a_0
tblib=1.7.0=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
typed-ast=1.4.3=py38h7f8727e_1
typeguard=2.13.0=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
voms=2.1.0rc0=h441846d_2
webencodings=0.5.1=py38_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.5.0=ha4553b6_1
